Seperate config file
