+++ .gitignore
@@ -3,3 +3,4 @@ package-lock.json
 node_modules
 desktop.ini
 yarn.lock
+db/config.js
