+++ db/config.template.js
@@ -1,0 +1,8 @@
+const config = {
+  host: 'localhost',
+  user: 'root',
+  password: 'root',
+  database: 'Systaurant'
+};
+
+export default config;
