+++ db/connectDB.js
@@ -1,11 +1,6 @@
 const mysql = require('mysql');
 
-const config = {
-  host: 'localhost',
-  user: 'root',
-  password: 'root',
-  database: 'Systaurant'
-};
+import config from './config';
 
 module.exports = function(callback) {
   const db_con = mysql.createConnection(config);
